pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gar-repro"
version = "0.1.0"
description = "Reproduction code for Generative Adversarial Reasoner (GAR)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "accelerate>=0.34.0",
  "datasets>=2.21.0",
  "numpy>=1.26.0",
  "openai>=1.51.0",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.2",
  "regex>=2024.7.24",
  "torch>=2.4.0",
  "transformers>=4.45.0",
  "trl>=0.11.4",
  "tqdm>=4.66.5",
]

[project.optional-dependencies]
# FULL_SCALE_REVERT: install this extra on Linux + NVIDIA GPU environments.
fullscale = [
  "vllm>=0.6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
